# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 천지연꽃신당 API
  version: 0.2.1
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/payments:
    post:
      summary: 결제 생성 (PENDING)
      security: [{ bearerAuth: [] }]
  /api/v1/payments/{id}:
    get:
      summary: 결제 조회
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
  /api/v1/payments/{id}/confirm:
    post:
      summary: 결제 승인 처리
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
  /api/v1/payments/{id}/cancel:
    post:
      summary: 결제 취소 처리
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
...